extract networking to a loader
