+++ app/src/main/java/com/moviefiend/torian/moviefiend/MovieListFragment.java
@@ -1,25 +1,31 @@
 package com.moviefiend.torian.moviefiend;
 
 import android.app.Fragment;
+import android.app.LoaderManager;
+import android.content.Loader;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
-import android.support.v7.widget.GridLayoutManager;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import com.android.volley.RequestQueue;
-import com.android.volley.Response;
-import com.android.volley.VolleyError;
-import com.android.volley.toolbox.Volley;
-import com.moviefiend.torian.moviefiend.network.GsonRequest;
 import com.moviefiend.torian.moviefiend.network.NowPlayingResponse;
 
-public class MovieListFragment extends Fragment {
+import java.util.ArrayList;
+
+public class MovieListFragment extends Fragment implements LoaderManager.LoaderCallbacks<ArrayList<NowPlayingResponse.MovieInfo>> {
+    public static final int MOVIE_LOADER_ID = 0;
     private RecyclerView mRecyclerView;
     private MovieAdapter mMovieAdapter;
+
+    @Override
+    public void onActivityCreated(Bundle savedInstanceState) {
+        super.onActivityCreated(savedInstanceState);
+
+        getLoaderManager().initLoader(MOVIE_LOADER_ID, null, this);
+    }
 
     @Nullable
     @Override
@@ -31,29 +37,23 @@ public class MovieListFragment extends Fragment {
         mMovieAdapter = new MovieAdapter();
         mRecyclerView.setAdapter(mMovieAdapter);
 
-        loadMovies();
-
         return view;
     }
 
-    private void loadMovies() {
-        RequestQueue queue = Volley.newRequestQueue(getActivity());
-        // TODO: Extract key to gradle/resource (can also extract the url... maybe)
-        String url ="http://api.themoviedb.org/3/movie/now_playing?api_key=4ef5ef8497673e7c432e66126be4a324";
-
-        GsonRequest<NowPlayingResponse> request = new GsonRequest<>(url, NowPlayingResponse.class, null,
-                new Response.Listener<NowPlayingResponse>() {
-                    @Override
-                    public void onResponse(NowPlayingResponse response) {
-                        mMovieAdapter.setMovies(response.getMovies());
-                    }
-                },
-                new Response.ErrorListener() {
-                    @Override
-                    public void onErrorResponse(VolleyError error) {
-                        // TODO: Error
-                    }
-                });
-        queue.add(request);
+//    <editor-fold Loader>
+    @Override
+    public Loader<ArrayList<NowPlayingResponse.MovieInfo>> onCreateLoader(int id, Bundle args) {
+        return new MovieInfoLoader(getActivity());
     }
+
+    @Override
+    public void onLoadFinished(Loader<ArrayList<NowPlayingResponse.MovieInfo>> loader, ArrayList<NowPlayingResponse.MovieInfo> data) {
+        mMovieAdapter.setMovies(data);
+    }
+
+    @Override
+    public void onLoaderReset(Loader<ArrayList<NowPlayingResponse.MovieInfo>> loader) {
+
+    }
+//    </editor-fold>
 }
